requirements.txt
flask
flask-cors
pymupdf
pdfplumber
pytesseract
Pillow
pandas
python-docx
werkzeug